# Cache read queries