--- a/src/cache/index.ts
+++ b/src/cache/index.ts
@@ -9,18 +9,42 @@ async function createCacheTable(dataSource?: DataSource) {
         "ttl" INTEGER NOT NULL,
         "query" TEXT UNIQUE NOT NULL,
         "results" TEXT
-    )`
+    );
+    `
 
     await dataSource?.internalConnection?.durableObject.executeQuery(statement, undefined, false)
 }
 
+function hasModifyingStatement(ast: any): boolean {
+    // Check if current node is a modifying statement
+    if (ast.type && ['insert', 'update', 'delete'].includes(ast.type.toLowerCase())) {
+        return true;
+    }
+
+    // Recursively check all properties of the AST
+    for (const key in ast) {
+        if (typeof ast[key] === 'object' && ast[key] !== null) {
+            if (Array.isArray(ast[key])) {
+                if (ast[key].some(item => hasModifyingStatement(item))) {
+                    return true;
+                }
+            } else if (hasModifyingStatement(ast[key])) {
+                return true;
+            }
+        }
+    }
+    return false;
+}
+
 export async function beforeQueryCache(sql: string, params?: any[], dataSource?: DataSource): Promise<any | null> {
-    if (dataSource?.source === Source.external && dataSource?.request.headers.has('X-Starbase-Cache')) {
+    let ast = parser.astify(sql);
+
+    if (!hasModifyingStatement(ast) && dataSource?.source === Source.external && dataSource?.request.headers.has('X-Starbase-Cache')) {
         // Currently we do not support caching queries that have dynamic parameters
         if (params?.length) return null
         await createCacheTable(dataSource)
-        const fetchCacheStatement = `SELECT timestamp, ttl, query, results FROM tmp_cache WHERE query='${sql}'`
-        const result = await dataSource.internalConnection?.durableObject.executeQuery(fetchCacheStatement, undefined, false) as any[];
+        const fetchCacheStatement = `SELECT timestamp, ttl, query, results FROM tmp_cache WHERE query = ?`
+        const result = await dataSource.internalConnection?.durableObject.executeQuery(fetchCacheStatement, [sql], false) as any[];
 
         if (result?.length) {
             const { timestamp, ttl, results } = result[0];
@@ -44,9 +68,9 @@ export async function afterQueryCache(sql: string, params: any[] | undefined, re
 
     try {
         let ast = parser.astify(sql);
-        const statementType = Array.isArray(ast) && ast.length ? ast[0].type : ast.type
-
-        if (statementType !== 'select' || 
+        
+        // If any modifying query exists within our SQL statement then we shouldn't proceed
+        if (hasModifyingStatement(ast) ||
             !(dataSource?.source === Source.external && dataSource?.request.headers.has('X-Starbase-Cache'))) return;
 
         const timestamp = Date.now();