--- a/src/operation.ts
+++ b/src/operation.ts
@@ -8,6 +8,9 @@ import { CloudflareD1Connection, MongoDBConnection, MySQLConnection, PostgreSQLC
 import { DataSource } from './types';
 import { Env } from './'
 import { MongoClient } from 'mongodb';
+import { afterQueryCache, beforeQueryCache } from './cache';
+
+let globalConnection: any = null;
 
 export type OperationQueueItem = {
     queries: { sql: string; params?: any[] }[];
@@ -46,7 +49,7 @@ async function afterQuery(sql: string, result: any, isRaw: boolean, dataSource?:
     result = isRaw ? transformRawResults(result, 'from') : result;
 
     // ## DO NOT REMOVE: POST QUERY HOOK ##
-    
+
     return isRaw ? transformRawResults(result, 'to') : result;
 }
 
@@ -131,6 +134,12 @@ export async function executeQuery(sql: string, params: any | undefined, isRaw: 
         return []
     }
 
+    // If a cached version of this query request exists, this function will fetch the cached results.
+    const cache = await beforeQueryCache(sql, params, dataSource)
+    if (cache) {
+        return cache
+    }
+
     const { sql: updatedSQL, params: updatedParams } = await beforeQuery(sql, params, dataSource, env)
     let response;
 
@@ -139,6 +148,9 @@ export async function executeQuery(sql: string, params: any | undefined, isRaw: 
     } else {
         response = await executeExternalQuery(updatedSQL, updatedParams, isRaw, dataSource, env);
     }
+
+    // If this is a cacheable query, this function will handle that logic.
+    await afterQueryCache(sql, updatedParams, response, dataSource)
 
     return await afterQuery(updatedSQL, response, isRaw, dataSource, env);
 }
@@ -248,30 +260,31 @@ export async function executeSDKQuery(sql: string, params: any | undefined, isRa
         return []
     }
     
-    let db;
-    
-    if (env?.EXTERNAL_DB_TYPE === 'postgres') {
-        const { database } = await createSDKPostgresConnection(env)
-        db = database
-    } else if (env?.EXTERNAL_DB_TYPE === 'mysql' && env) {
-        const { database } = await createSDKMySQLConnection(env)
-        db = database
-    } else if (env?.EXTERNAL_DB_TYPE === 'mongo' && env) {
-        const { database } = await createSDKMongoConnection(env)
-        db = database
-    } else if (env?.EXTERNAL_DB_TYPE === 'sqlite' && env?.EXTERNAL_DB_CLOUDFLARE_API_KEY && env) {
-        const { database } = await createSDKCloudflareConnection(env)
-        db = database
-    } else if (env?.EXTERNAL_DB_TYPE === 'sqlite' && env?.EXTERNAL_DB_STARBASEDB_URI && env) {
-        const { database } = await createSDKStarbaseConnection(env)
-        db = database
-    } else if (env?.EXTERNAL_DB_TYPE === 'sqlite' && env?.EXTERNAL_DB_TURSO_URI && env) {
-        const { database } = await createSDKTursoConnection(env)
-        db = database
+    // Initialize connection if it doesn't exist
+    if (!globalConnection) {
+        if (env?.EXTERNAL_DB_TYPE === 'postgres') {
+            const { database } = await createSDKPostgresConnection(env)
+            globalConnection = database;
+        } else if (env?.EXTERNAL_DB_TYPE === 'mysql' && env) {
+            const { database } = await createSDKMySQLConnection(env)
+            globalConnection = database;
+        } else if (env?.EXTERNAL_DB_TYPE === 'mongo' && env) {
+            const { database } = await createSDKMongoConnection(env)
+            globalConnection = database;
+        } else if (env?.EXTERNAL_DB_TYPE === 'sqlite' && env?.EXTERNAL_DB_CLOUDFLARE_API_KEY && env) {
+            const { database } = await createSDKCloudflareConnection(env)
+            globalConnection = database;
+        } else if (env?.EXTERNAL_DB_TYPE === 'sqlite' && env?.EXTERNAL_DB_STARBASEDB_URI && env) {
+            const { database } = await createSDKStarbaseConnection(env)
+            globalConnection = database;
+        } else if (env?.EXTERNAL_DB_TYPE === 'sqlite' && env?.EXTERNAL_DB_TURSO_URI && env) {
+            const { database } = await createSDKTursoConnection(env)
+            globalConnection = database;
+        }
+        
+        await globalConnection.connect();
     }
 
-    await db.connect();
-    const { data } = await db.raw(sql, params);
-    
+    const { data } = await globalConnection.raw(sql, params);
     return data;
 }